_OpenGL_ *FIX* - Re-added support for OpenGL 2.1 shaders (GLSL 1.0, 1.1, and 1.2).

+++ Source/Engine/OpenGL/Include/_OpenGL.h
@@ -47,7 +47,7 @@ uint glGetInteger(uint id);
 // One function to bind all function pointers
 //==========================================================================================================
 
-bool InitOpenGL (float requiredVersion);
+bool InitOpenGL (uint requiredVersion);
 
 //==========================================================================================================
 // Handy macro to check for an OpenGL error
